+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.71] - 16/09/2026
 
+### Added
+
 - Add `extract_ner` pattern for LLM
 - Add spaCy 'Named Entity Recognition' plugin: `atlas -u ner`
 - Add General-purpose `atlas -u converter` plugin with support for HH:MM:SS-to-Seconds and vice versa
 - Add `--format=srt` option to YouTube transcript downloader (`--youtube`)
+
+### Fixed
+
+- Fix context window error by stopping LLM inference if input context window is larger than the default setting
 
 ## [0.1.7] - 11/09/2026